Code optimization and bug fix
onBackground even will onty return data if the new battery usage is different than the latest one in the buffer.
onBackground will half the data buffer if it's too big to be sent, repeating as necessary until it can successfully sent.
analyzeAndStoreData check if each battery levels are newer before adding to history.