Changed when we read and store of History
History view is read at getGlanceView and getMainView as well as onBackgroundData (but only if it's null there). The background process does NOT read it.
History view is stored at onStop and only if it has changed as well as at the end of onBackgroundData only if we read more than two array elements.
The background process is stopped while running the Glance or Main View
The Glance view has its own timer and can set the history array.